backend/requirements.txt
fastapi[standard]
uvicorn
pymongo
python-dotenv
pwdlib[argon2]
pyjwt
python-multipart
redis